Updated task controller
Added role-based access control for task management actions.

+++ app/Http/Controllers/TaskController.php
@@ -11,14 +11,24 @@ use App\Models\AAUser;
 class TaskController extends Controller
 {
     public function index() {
+        if(Auth::user()->role=== 'admin'){
+            $tasks = Task::all();
+        } else{
         $tasks = Task::where('user_id', Auth::id())->get();
+        }
         return view('tasks.index', compact('tasks'));
     }
 
     public function create() {
-        return view('tasks.create');
+        if(Auth::user()->role==='guest'){
+        return redirect()->route('tasks.index')
+            ->with('error' , 'Guests cannot create tasks');
+        }
+        
+        
          $categories= Category::all();
-        return view('tasks.create', compact('categories'));
+        $users = AAUser::all();
+        return view('tasks.create', compact('categories' , 'users'));
     }
 
     public function store(Request $request) {
@@ -43,7 +53,10 @@ class TaskController extends Controller
         return redirect()->route('tasks.index');
     }
     public function edit(Task $task) {
-    return view('tasks.edit', compact('task'));
+        if(Auth::user()->role==='guest'){
+            return redirect()->route('task.index')
+                ->with('error', 'Guests cannot edit tasks');
+                }
         $categories = Category::all();
         return view ('tasks.edit', compact ('task', 'categories'));
 }
@@ -67,6 +80,10 @@ public function update(Request $request, Task $task) {
     return redirect()->route('tasks.index');
 }
 public function destroy(Task $task) {
+    if(Auth::user()->role==='guest'){
+        return redirect()->route('task.index')
+            ->with('error', 'Guests cannot delete tasks');
+    }
     $task->delete();
     return redirect()->route('tasks.index');
 }
